requirements/base.txt
# Bleeding edge Django
django==1.8.3

# Configuration
django-environ==0.3.0
django-secure==1.0.1



# Forms
django-braces==1.8.0
django-crispy-forms==1.4.0
django-floppyforms==1.4.1

# Models
django-model-utils==2.2

# Images
Pillow==2.9.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.21.0

# For the persistence stores
mysqlclient==1.3.6


# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.8.0

# Time zones support
pytz==2015.4



# Your custom requirements go here
django-filters==0.2.1
django-autocomplete-light==2.2.4
https://github.com/ad-m/django-atom/archive/v0.1.0.zip
